Extract editor selection update logic to custom hook
